judicary_backend/requirements.txt
# ---- Core ----
flask==3.0.0
flask-cors==4.0.0
flask-mongoengine==1.0.0
flask-jwt-extended==4.6.0
gunicorn==21.2.0
mongoengine==0.27.0
pymongo>=4.6.1
python-dotenv==1.0.0
bcrypt>=4.1.2
PyJWT>=2.8.0
marshmallow==3.20.1

# ---- HTTP / Scraping ----
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.1.0
selenium>=4.16.0
fake-useragent==1.4.0
cloudscraper>=1.2.71

# ---- Scheduling ----
apscheduler==3.10.4

# ---- Document parsing ----
pdfplumber==0.10.3
python-dateutil==2.8.2

# ---- AI ----
google-generativeai>=0.8.0
